+++ app/mailers/devise_sendgrid_mailer.rb
@@ -1,0 +1,24 @@
+class DeviseSendgridMailer < Devise::Mailer
+  include ApplicationHelper
+  include Devise::Controllers::UrlHelpers
+  include Rails.application.routes.url_helpers
+
+  def confirmation_instructions(record, token, opts={})
+    if record.instance_of? User
+      link_url = api_ui_v1_user_confirmation_path(confirmation_token: token)
+      SendgridMailer.account_confirmation(record, link_url: link_url).deliver
+    else
+      super
+    end
+  end
+
+  def reset_password_instructions(record, token, opts={})
+    if record.instance_of? User
+      link_url = edit_api_ui_v1_user_password_path(reset_password_token: token)
+      SendgridMailer.reset_password(record, link_url: link_url).deliver
+    else
+      super
+    end
+  end
+
+end