+++ app/models/system_log.rb
@@ -1,0 +1,3 @@
+class SystemLog < ApplicationRecord
+
+end