+++ db/migrate/20200201223219_create_system_logs.rb
@@ -1,0 +1,10 @@
+class CreateSystemLogs < ActiveRecord::Migration[6.0]
+  def change
+    create_table :system_logs do |t|
+      t.integer :log_level, limit: 2
+      t.string :message
+
+      t.timestamps
+    end
+  end
+end